Fix the misterious failure of building on TravisCI
v0.4-0 was failing the checks on travis for R >= 3.6.0, despite passing
all the checks on CRAN. Further digging revealed that 'devtools::test()'
was also failing, but only for the first run in a session. Subsequent
runs in the same session passed.

Clearing the values saved by 'devtools::test()' was not
helping. Moreover, depending on whether the values were recreated by the
first run in the session or by a subsequent run, the state saved in the
object was either "Rejection" or "Rounding". The net result being that
could happen also that the first run ws ok but all subsequent runs
failed (in this scenario 'R CMD check' was also failing).

The error was occuring in 'test-garch.R' but the cause turned out to be
a leftover 'suppressWarnings(RNGversion("3.5.0"))' in 'test-VaR' which
was causing the sample.kind (the third element of the value of
'RNGkind()') to switch from "Rejection" to "Rounding". This explains the
different result from subsequent runs of 'devtools::test()' in the same
session.

Just deleting the offending 'suppressWarnings(RNGversion("3.5.0"))' and
recreating 'a.RDS' and 'a_pred.RDS' resolved the problem.

In the process I streamlined somewhat the handling of seeds for the
random number generators.